+++ packages/rtmpstreamer/lib/rtmpstreamer.dart
@@ -1,4 +1,2 @@
-library rtmpstreamer;
-
 export 'src/rtmp_camera_preview.dart';
 export 'src/rtmp_streamer_controller.dart';