+++ packages/rtmpstreamer/lib/src/preview/rtmp_camera_preview_native.dart
@@ -1,0 +1,15 @@
+import 'dart:io';
+
+import 'package:flutter/widgets.dart';
+
+class RtmpCameraPreview extends StatelessWidget {
+  const RtmpCameraPreview({super.key});
+
+  static const viewType = 'rtmpstreamer_camera_preview';
+
+  @override
+  Widget build(BuildContext context) {
+    if (!Platform.isAndroid) return const SizedBox.shrink();
+    return const AndroidView(viewType: viewType);
+  }
+}