+++ packages/rtmpstreamer/lib/src/rtmp_camera_preview.dart
@@ -1,18 +1,3 @@
-import 'dart:io';
-
-import 'package:flutter/foundation.dart';
-import 'package:flutter/widgets.dart';
-
-class RtmpCameraPreview extends StatelessWidget {
-  const RtmpCameraPreview({super.key});
-
-  static const viewType = 'rtmpstreamer_camera_preview';
-
-  @override
-  Widget build(BuildContext context) {
-    if (kIsWeb || !Platform.isAndroid) {
-      return const SizedBox.shrink();
-    }
-    return const AndroidView(viewType: viewType);
-  }
-}
+export 'preview/rtmp_camera_preview_stub.dart'
+    if (dart.library.io) 'preview/rtmp_camera_preview_native.dart'
+    if (dart.library.html) 'preview/rtmp_camera_preview_web.dart';